+++ src/pages/Home/Home.tsx
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import { Carousel } from '../../components/Home/Carousel';
 import { DailyBrief } from '../../components/Home/DailyBrief';
 import { Shortcut } from '../../components/Home/Shortcut';
@@ -5,11 +6,15 @@ import { TodoList } from '../../components/Home/TodoList';
 import { HomeContainer } from '../../styles/Home/HomeSC';
 
 export function Home(): JSX.Element {
-  // set done, todo를 todo에서 가져오기
+  const [todo, setTodo] = useState<Array<[string, number, string]>>([]); //  할 일 목록
+  const [done, setDone] = useState<Array<[string, number, string]>>([]); //  완료된 일 목록
+  const [TabMenu, setTabMenu] = useState([1, 0]); //  탭 전환 상태값
+  const [tglEditTodo, setTglEditTodo] = useState<boolean[]>([]); //  할일 수정 상태 표출 / 숨기기
+
   return (
     <HomeContainer>
       <Carousel />
-      <DailyBrief />
+      <DailyBrief todo={todo} done={done} />
       <div
         style={{
           display: 'flex',
@@ -18,8 +23,22 @@ export function Home(): JSX.Element {
           marginTop: '24px',
         }}
       >
-        <Shortcut></Shortcut>
-        <TodoList></TodoList>
+        <Shortcut //
+          todo={todo}
+          setTodo={setTodo}
+          setTabMenu={setTabMenu}
+          setTglEditTodo={setTglEditTodo}
+        ></Shortcut>
+        <TodoList
+          todo={todo}
+          setTodo={setTodo}
+          done={done}
+          setDone={setDone}
+          TabMenu={TabMenu}
+          setTabMenu={setTabMenu}
+          tglEditTodo={tglEditTodo}
+          setTglEditTodo={setTglEditTodo}
+        ></TodoList>
       </div>
     </HomeContainer>
   );
