+++ src/styles/Home/ShortcutGallerySC.tsx
@@ -1,0 +1,14 @@
+import { styled } from 'styled-components';
+// import { theme } from '../Theme';
+
+export const ImgsContainer = styled.div`
+  width: 100%;
+  backgroung-color: red;
+  height: 200px;
+`;
+
+export const ImgItems = styled.div`
+  width: 100px;
+  height: 100px;
+  background-size: cover;
+`;
